+++ pkg/storage/stores/shipper/compactor/metrics.go
@@ -14,6 +14,7 @@ type metrics struct {
 	compactTablesOperationTotal           *prometheus.CounterVec
 	compactTablesOperationDurationSeconds prometheus.Gauge
 	compactTablesOperationLastSuccess     prometheus.Gauge
+	compactorRunning                      prometheus.Gauge
 }
 
 func newMetrics(r prometheus.Registerer) *metrics {
@@ -32,6 +33,11 @@ func newMetrics(r prometheus.Registerer) *metrics {
 			Namespace: "loki_boltdb_shipper",
 			Name:      "compact_tables_operation_last_successful_run_timestamp_seconds",
 			Help:      "Unix timestamp of the last successful compaction run",
+		}),
+		compactorRunning: promauto.With(r).NewGauge(prometheus.GaugeOpts{
+			Namespace: "loki_boltdb_shipper",
+			Name:      "compactor_running",
+			Help:      "Value will be 1 if compactor is currently running on this instance",
 		}),
 	}